[SPARK-36868][SQL][FOLLOWUP] move paring Create Function from catalyst to core
