+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/v2Commands.scala
@@ -688,17 +688,6 @@ case class DescribeFunction(child: LogicalPlan, isExtended: Boolean) extends Una
 }
 
 /**
- * The logical plan of the CREATE TEMPORARY FUNCTION command.
- */
-case class CreateTempFunction(
-    nameParts: Seq[String],
-    className: String,
-    resources: Seq[FunctionResource],
-    ifExists: Boolean,
-    replace: Boolean) extends LeafCommand {
-}
-
-/**
  * The logical plan of the CREATE FUNCTION command.
  */
 case class CreateFunction(
