+++ sql/core/src/main/scala/org/apache/spark/sql/catalyst/analysis/ResolveSessionCatalog.scala
@@ -19,7 +19,7 @@ package org.apache.spark.sql.catalyst.analysis
 
 import org.apache.spark.sql.SaveMode
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.catalyst.catalog.{BucketSpec, CatalogStorageFormat, CatalogTable, CatalogTableType, CatalogUtils, FunctionResource}
+import org.apache.spark.sql.catalyst.catalog.{BucketSpec, CatalogStorageFormat, CatalogTable, CatalogTableType, CatalogUtils}
 import org.apache.spark.sql.catalyst.expressions.{Alias, Attribute}
 import org.apache.spark.sql.catalyst.plans.logical._
 import org.apache.spark.sql.catalyst.rules.Rule
@@ -447,14 +447,24 @@ class ResolveSessionCatalog(val catalogManager: CatalogManager)
       val funcIdentifier = identifier.asFunctionIdentifier
       DropFunctionCommand(funcIdentifier.database, funcIdentifier.funcName, ifExists, isTemp)
 
-    case CreateTempFunction(nameParts, className, resources, ignoreIfExists, replace) =>
-      // temp func doesn't belong to any catalog and we shouldn't resolve catalog in the name.
-      convertToV1CreateFunction(nameParts, className, resources, true, ignoreIfExists, replace)
-
     case CreateFunction(ResolvedDBObjectName(catalog, nameParts),
         className, resources, ignoreIfExists, replace) =>
       if (isSessionCatalog(catalog)) {
-        convertToV1CreateFunction(nameParts, className, resources, false, ignoreIfExists, replace)
+        val database = if (nameParts.length > 2) {
+          throw QueryCompilationErrors.unsupportedFunctionNameError(nameParts.quoted)
+        } else if (nameParts.length == 2) {
+          Some(nameParts.head)
+        } else {
+          None
+        }
+        CreateFunctionCommand(
+          database,
+          nameParts.last,
+          className,
+          resources,
+          false,
+          ignoreIfExists,
+          replace)
       } else {
         throw QueryCompilationErrors.functionUnsupportedInV2CatalogError()
       }
@@ -463,35 +473,6 @@ class ResolveSessionCatalog(val catalogManager: CatalogManager)
       // Fallback to v1 command
       val funcIdentifier = identifier.asFunctionIdentifier
       RefreshFunctionCommand(funcIdentifier.database, funcIdentifier.funcName)
-  }
-
-  private def convertToV1CreateFunction(
-      nameParts: Seq[String],
-      className: String,
-      resources: Seq[FunctionResource],
-      isTemp: Boolean,
-      ignoreIfExists: Boolean,
-      replace: Boolean) = {
-    val database = if (nameParts.length > 2) {
-      throw QueryCompilationErrors.unsupportedFunctionNameError(nameParts.quoted)
-    } else if (nameParts.length == 2) {
-      Some(nameParts.head)
-    } else {
-      None
-    }
-    CreateFunctionCommand(
-      database,
-      nameParts.last,
-      className,
-      resources,
-      isTemp,
-      ignoreIfExists,
-      replace)
-  }
-
-  private def parseV1Table(tableName: Seq[String], sql: String): Seq[String] = tableName match {
-    case SessionCatalogAndTable(_, tbl) => tbl
-    case _ => throw QueryCompilationErrors.sqlOnlySupportedWithV1TablesError(sql)
   }
 
   private def getStorageFormatAndProvider(
